Move pkg/utils/kubernetes/health.CheckTunnelConnection to pkg/operation/botanist.CheckTunnelConnection
